Fix x64-leak test isolation + uint32 Optuna seed + drop --timeout=300
Finding: Three CI failures on PR #72 fix-emit-script-x64-fidelity:
  1. TestAutoDispatch tests fail with 'Seed must be between 0 and 2**32 - 1'
     because jax.random.bits() returns uint64 when x64 is enabled process-globally.
     The x64 flag is leaked by test_run_recipe_to_idata_laplace_multiphase_uses_dense_imm
     (slow test, in-process run_recipe_to_idata with gp_regression requires_x64=True).
  2. test-e2e CI workflow fails at startup: 'unrecognized arguments: --timeout=300'
     because pytest-timeout is not in the dev deps.

Fix:
  1. tune.py:748 -- dtype=jnp.uint32 forces Optuna seed to always be uint32
     regardless of x64 config state (genuine latent bug, not only a test bug).
  2. tests/conftest.py -- autouse fixture saves jax_enable_x64 before each test
     and restores it after, preventing any in-process test from leaking x64 to
     subsequent tests in the same session.
  3. .github/workflows/test-e2e.yml -- drop --timeout=300 (pytest-timeout not
     in deps; job-level timeout-minutes=90 + -n 1 already bound it adequately).

+++ tuningfork/calibration/tune.py
@@ -743,7 +743,9 @@ def tune_algorithm(
     # ------------------------------------------------------------------
     # 3. Optuna study with selected sampler
     # ------------------------------------------------------------------
-    sampler_seed = int(jax.random.bits(jax.random.fold_in(rng_key_study, 0)))
+    sampler_seed = int(
+        jax.random.bits(jax.random.fold_in(rng_key_study, 0), dtype=jnp.uint32)
+    )
     if sampler == "tpe":
         optuna_sampler = optuna.samplers.TPESampler(seed=sampler_seed)
     elif sampler == "random":
